Check for the SYN flag only
The filter originally matched only packets where only the SYN flag was
set. If other flags are present (i.e. the ECE flag), the filter would
not match.

This patch ensures that the filter matches as soon as the SYN flag is
set, no matter whether any other flags are set.

Close #59.